fetchData, deleteClass, uptdateClass
fetchData: fetchData: fetch data when user  click  in "ver aulas" on screen "CriarCategorias".

deleteClass: delete a class when the user click on trash icon .

updateClass: not working, but substitut the data with new value.

modal: the modal has been created to set new values when use updateClass.
